+++ CHANGELOG.md
@@ -8,6 +8,9 @@
 ### Glean
 - Updated to v58.0.0 ([#6150](https://github.com/mozilla/application-services/pull/6150))
 
+### Suggest
+- Improved full keyword display for AMP suggestions
+
 [Full Changelog](In progress)
 
 # v124.0 (_2024-02-15_)